requirements.txt
discord.py
spotipy
python-dotenv
yt-dlp
PyNaCl